Move the IgnoredManagedGroupMembership model into auditor app
Move the admin and tests as well. Temporarily update import statements
in the main app until everything else can be moved. Many tests are
still broken.
